[PW-6532] Declarative data patch for 8.0.0
Dropping patch for v2.4.4 since that version is not officially supported.
